backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-docx==1.1.2
pydantic==2.10.4
pydantic-settings==2.7.1
websockets==14.1
aiofiles==24.1.0
anthropic==0.42.0
